pyproject.toml
[tool.poetry]
name = "ml-utils"
version = "0.1.0"
description = ""
authors = ["Bogdan Gulowaty <b.gulowaty@gmail.com>"]
readme = "README.md"
packages = [{include = "mlutils"}]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.24.2"
mlxtend = "^0.21.0"
scikit-learn = ">=1.3"
python-box = "^7.0.1"
pandas = "*"
loguru = "*"
mlflow = "^2.3.0"
joblib = "^1.2.0"
tqdm = "^4.65.0"
papermill = "2.5.0"
ray = {extras = ["default"], version = "^2.4.0"}
pyyaml = "*"
cachetools = "^5.3.1"
xxhash = "^3.4.1"
numba = "*"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
cvxpy = "^1.3.0"
pypruning = {git = "https://github.com/sbuschjaeger/PyPruning.git"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
